+++ movies_app/lib/UI Screens/HomeScreen/Screen/HomeScreenfinal.dart
@@ -13,14 +13,13 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-  int _currentIndex = 0; // Track the current selected index
+  int _currentIndex = 0; 
   
-  // List of screens to navigate to (you'll need to create these)
   final List<Widget> _screens = [
-    HomeContent(), // This will contain your current home content
-    SearchScreen(), // You'll need to create this
-    MovieListScreen(), // You'll need to create this
-    MovieListScreen(), // You'll need to create this
+    HomeContent(),
+    SearchScreen(), 
+    MovieListScreen(),
+    MovieListScreen(), 
   ];
   // SearchScreen
 
@@ -102,98 +101,3 @@ class HomeContent extends StatelessWidget {
     );
   }
 }
-
-// // Placeholder for other screens - you'll need to implement these
-// class SearchScreen extends StatelessWidget {
-//   @override
-//   Widget build(BuildContext context) {
-//     return Center(child: Text('Search Screen'));
-//   }
-// }
-
-// class BrowseScreen extends StatelessWidget {
-//   @override
-//   Widget build(BuildContext context) {
-//     return Center(child: Text('Browse Screen'));
-//   }
-// }
-
-// class ProfileScreen extends StatelessWidget {
-//   @override
-//   Widget build(BuildContext context) {
-//     return Center(child: Text('Profile Screen'));
-//   }
-// }
-// import 'package:flutter/material.dart';
-// import 'package:flutter_bloc/flutter_bloc.dart';
-// import 'package:movies_app/Domain/Models/movie.dart';
-// import 'package:movies_app/UI%20Screens/BrowserScreen/Cubit/movie_cubit.dart';
-// import 'package:movies_app/UI%20Screens/HomeScreen/widgets/MoviesCategoryHomeWidget.dart';
-// import 'package:movies_app/UI%20Screens/HomeScreen/widgets/SliderContainerCustom.dart';
-
-// class HomeScreen extends StatefulWidget {
-//   @override
-//   _HomeScreenState createState() => _HomeScreenState();
-// }
-
-// class _HomeScreenState extends State<HomeScreen> {
-//   @override
-//   Widget build(BuildContext context) {
-//     final movieCubit = BlocProvider.of<MovieCubit>(context);
-//     movieCubit.loadMovies();
-
-//     return Scaffold(
-//       backgroundColor: Colors.black,
-//       body: BlocBuilder<MovieCubit, List<Movie>>(
-//         builder: (context, movies) {
-//           if (movies.isEmpty) {
-//             return Center(child: CircularProgressIndicator());
-//           }
-
-//           final featuredMovies = movies.take(5).toList();
-
-//           List<Movie> getMoviesByCategory(String category) {
-//             return movies
-//                 .where((movie) => movie.genres?.contains(category) ?? false)
-//                 .toList();
-//           }
-
-//           return SingleChildScrollView(
-//             child: Column(
-//               crossAxisAlignment: CrossAxisAlignment.start,
-//               children: [
-//                 SliderContainerCustom(featuredMovies: featuredMovies),
-//                 MoviesCategoryHomeWidget(
-//                   context: context,
-//                   title: "Action",
-//                   movies: getMoviesByCategory("Action"),
-//                 ),
-//                 MoviesCategoryHomeWidget(
-//                   context: context,
-//                   title: "Drama",
-//                   movies: getMoviesByCategory("Drama"),
-//                 ),
-//                 MoviesCategoryHomeWidget(
-//                   context: context,
-//                   title: "Comedy",
-//                   movies: getMoviesByCategory("Comedy"),
-//                 ),
-//               ],
-//             ),
-//           );
-//         },
-//       ),
-//       bottomNavigationBar: BottomNavigationBar(
-//         backgroundColor: Colors.black,
-//         selectedItemColor: Colors.yellow,
-//         unselectedItemColor: Colors.white54,
-//         items: [
-//           BottomNavigationBarItem(icon: Icon(Icons.home), label: ""),
-//           BottomNavigationBarItem(icon: Icon(Icons.search), label: ""),
-//           // BottomNavigationBarItem(icon: Icon(Icons.compass_calibration_outlined), label: ""),
-//           BottomNavigationBarItem(icon: Icon(Icons.person), label: ""),
-//         ],
-//       ),
-//     );
-//   }
-// }
