+++ movies_app/lib/oldstuff/MovieDetailsScreen/model for test/model.dart
